refactor: extract ExecutionArgs types to separate file
