+++ src/execution/hooks.ts
@@ -1,19 +1,10 @@
 /** @category Execution */
 
 import type { SharedExecutionContext } from './createSharedExecutionContext.ts';
-import type { ValidatedExecutionArgs } from './Executor.ts';
-
-/** Information passed to hooks after asynchronous execution work has finished. */
-export interface AsyncWorkFinishedInfo {
-  /** Validated execution arguments for the operation that finished async work. */
-  validatedExecutionArgs: ValidatedExecutionArgs;
-}
-
-/** Optional hooks invoked during GraphQL execution. */
-export interface ExecutionHooks {
-  /** Called after all tracked asynchronous execution work has settled. */
-  asyncWorkFinished?: (info: AsyncWorkFinishedInfo) => void;
-}
+import type {
+  AsyncWorkFinishedInfo,
+  ValidatedExecutionArgs,
+} from './ExecutionArgs.ts';
 
 function runHookSafely<TInfo>(hook: (info: TInfo) => void, info: TInfo): void {
   try {
